impl grid

+++ examples/glfw_imgui/scene.h
@@ -10,6 +10,6 @@ class Scene {
 public:
   Scene();
   ~Scene();
-  void Render(int width, int height, const float clear_color[4],
+  void Render(float width, float height, const float clear_color[4],
               const rectray::MouseState &mouse);
 };
